% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{check assertion}
\usage{
check(..., env = parent.frame())
}
\arguments{
\item{...}{character strings}

\item{env}{environment}
}
\value{
\code{invisible(NULL)}
}
\description{
check assertion
}
