% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSS_helper_functions.R
\name{check_tp}
\alias{check_tp}
\title{Check Visium tissue positions}
\usage{
check_tp(
  tp,
  L_capture_area = NULL,
  tp_columns = c("barcode", "in_tissue"),
  array_columns = c("array_row", "array_col"),
  pixel_columns = c("pxl_row_in_fullres", "pxl_col_in_fullres")
)
}
\description{
Check Visium tissue positions
}
\keyword{internal}
