% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sfe_plot.R
\name{create_sfe_plot}
\alias{create_sfe_plot}
\title{Plot a spatially aligned Visium image and spot geometry}
\usage{
create_sfe_plot(
  sfe,
  image_resolution,
  geom_id = "spatial_pixels",
  plot_title = "Spatially aligned Pixel Positions",
  color_by = NULL,
  colors_use = NULL,
  metadata_slot = "spot_metadata",
  color_source = NULL
)
}
\arguments{
\item{sfe}{A \code{SpatialFeatureExperiment} containing the requested image
and column geometry.}

\item{image_resolution}{Character scalar. Image identifier passed to
\code{\link[SpatialExperiment]{getImg}}, for example \code{"lowres"} or
\code{"hires"}.}

\item{geom_id}{Character scaler. Name of the spatially resolved column
geometry retrieved with \code{\link[SpatialFeatureExperiment]{colGeometry}}.
Available options include \code{"spatial_array"} and \code{"spatial_pixels"}.}

\item{plot_title}{Character scalar or \code{NULL}. Plot title to be used.}

\item{color_by}{Character scalar or \code{NULL}. Name of a metadata column
or the selected geometry field used to color spot outlines. If \code{NULL},
all spots are drawn using one fixed color.}

\item{colors_use}{Character vector or \code{NULL}. Colors used for spot
outlines. When \code{color_by = NULL}, this must contain one color. When
\code{color_by} is supplied, it may be either an unnamed vector with one
color per observed value or a named vector whose names match the observed
values. If insufficient colors are supplied, a default qualitative palette
is generated.
@param metadata_slot Optional, Character Scalar. Name of the metadata slot
to look for the \code{color_by} column, if \code{metadata} is selected as \code{color_source}.}

\item{colors_source}{Character vector or \code{NULL} (default). When
\code{color_by} is not \code{NULL}, this must contain the slot in the sfe object,
where the color information can be retrieved. Available options include \code{geometry}
and \code{metadata}.}
}
\value{
A \code{ggplot} object for the selected \code{geom_id}
containing the transformed image raster and the specific spatial spot geometry.

Missing values in \code{color_by} are displayed transparently.
}
\description{
Create a \code{ggplot2} overlay of a spatially transformed Visium image and
an \code{sf} spot geometry stored in a
\code{SpatialFeatureExperiment}. Spots can be displayed using one fixed
outline color or colored according to a geometry metadata column.
}
\examples{
\dontrun{
p <- create_sfe_plot(
  sfe = sfe,
  image_resolution = "lowres"
)

p_selected <- create_sfe_plot(
  sfe = sfe,
  image_resolution = "lowres",
  color_by = "selected",
  colors_use = c(
    selected = "cornflowerblue",
    out = "lightgrey"
  ),
  
  plot_title = "Selected spots"
)
}

}
\seealso{
\code{\link{load_sfe}},
\code{\link{raster_to_ggdf}},
}
