% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_vignette_files.R
\name{download_vignette_files}
\alias{download_vignette_files}
\title{Download vignette files}
\usage{
download_vignette_files(
  output_dir,
  visium_sample = c("Visium_FFPE_Human_Ovarian_Cancer",
    "CytAssist_11mm_FFPE_Human_Ovarian_Carcinoma"),
  matrix_file_format = "h5",
  matrix_name = "filtered",
  download_handler = NULL,
  image_resolution = "lowres"
)
}
\arguments{
\item{output_dir}{Parent output directory, where a folder for each visium
sample, matching the expected Space Ranger output structure, will be created.}

\item{visium_sample}{Character vector. Supported samples are
\describe{
\item{\code{Visium_FFPE_Human_Ovarian_Cancer}}{Ovarian Cancer sample
generated using Space Ranger Version 1.3.0 with an Array Size of 6.5mm.}
\item{\code{CytAssist_11mm_FFPE_Human_Ovarian_Carcinoma}}{Ovarian Cancer
Sample generated using Space Ranger Version 2.0.0 with an Array Size of 11mm.}}
The sample name  \code{visium_sample} has to match the file name from
the \verb{Batch download} command as specified on the Visium dataset download site with
\verb{https://cf.10xgenomics.com/samples/spatial-exp/<spaceranger_version>/<visium_sample>}.
@param spaceranger_version Optional, named character vector. Spaceranger
version used to generate the \code{visium_sample} file. Defaults to NULL, since
the SpaceRanger versions are resolved automatically for the vignette data.}

\item{matrix_file_format}{Character scalar. Defines the file format of the
matrix, which will be downloaded. Available options are \code{"h5"} for
downloading the \code{h5} file or \code{"sparse"} for the \code{tar.gz} file. The
compressed \code{tar.gz} file will be unpacked after the successful download,
and the \code{tar.gz} file will be removed, thereafter.}

\item{matrix_name}{Character vector specifying one or both matrices to
download. Available values are \code{"filtered"} (default) and \code{"raw"}.}

\item{download_handler}{Optional, character scalar. Download handler
used for downloading the requested files. Supported methods are \code{curl},
\code{libcurl}, \code{wget}, \code{auto}. If \code{NULL} (default) the currently selected
download handler will be used.}

\item{image_resolution}{Character vector specifying one or both tissue-image
resolutions to retain after extraction. Available options are \code{lowres}
for the low resolution image (default) and \code{hires} for the
high resolution image.}
}
\value{
Output directory where the downloaded files and folders can be found,
matching the Space Ranger output structure \verb{<visium_dir>/<visium_sample>/outs/}
}
\description{
Download vignette files
}
\examples{
\dontrun{
# Download Human Ovarian Cancer 6.5mm & Human Ovarian Cancer 11mm datasets
 with low & high resolution image
download_vignette_files(
  output_dir = "path/to/output/directory",
  visium_sample = c("Visium_FFPE_Human_Ovarian_Cancer", 
  "CytAssist_11mm_FFPE_Human_Ovarian_Carcinoma"),
  image_resolution=c("lowres","hires")
)

# Download Human Ovarian Cancer 11mm dataset with low resolution image
download_vignette_files(
  output_dir = "path/to/output/directory",
  visium_sample =  "CytAssist_11mm_FFPE_Human_Ovarian_Carcinoma"
)
}

}
