% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_counts_per_spot.R
\name{get_cps}
\alias{get_cps}
\title{Obtain counts per Visium spot}
\usage{
get_cps(
  visium_dir,
  visium_sample,
  matrix_name = c("filtered", "raw"),
  exclude_deprecated_genes = TRUE,
  remove_background_spots = FALSE
)
}
\arguments{
\item{visium_dir}{Character scalar. Base directory containing the Visium
sample directories.}

\item{visium_sample}{Character scalar. Name or identifier of the Visium
sample directory.}

\item{matrix_name}{Character scaler. Name of the matrix used for calculating
the counts per spot. Either \code{raw} or \code{filtered}.}

\item{exclude_deprecated_genes}{Logical. Whether features marked as
\code{DEPRECATED} should be excluded from the calculation. Defaults to \code{TRUE}.}

\item{remove_background_spots}{Logical. Whether background spots should
be removed from the returned object. Defaults to \code{FALSE}, so background spots
are retained and can be compared to tissue positions. Selecting \code{FALSE} requires
the raw matrix to be available, while \code{TRUE} enables using the filtered matrix,
only.}
}
\description{
This function is intended as a diagnostic step for checking whether high
counts per spot match the tissue position.
}
