% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spex.R
\name{load_spex}
\alias{load_spex}
\title{Load a \code{SpatialExperiment} Object for a 10x Genomics Visium sample}
\usage{
load_spex(
  visium_dir,
  visium_sample,
  matrix_name = "filtered",
  matrix_file_format = c("h5", "sparse"),
  image_resolution = c("lowres", "hires")
)
}
\arguments{
\item{visium_dir}{Character scalar. Base directory containing the Visium
sample directories.}

\item{visium_sample}{Character scalar. Name of the Visium sample directory.}

\item{matrix_name}{Character scalar. Name of Matrix passed to
\code{\link[SpatialExperiment]{read10xVisium}} through its
\code{data} argument. Available options are \code{"filtered"} (default)
or \code{"raw"}.}

\item{matrix_file_format}{Character scalar. Matrix type used for the \code{type}
argument in \code{\link[SpatialExperiment]{read10xVisium}},
either \code{"h5"} or \code{"sparse"}.}

\item{image_resolution}{Character scalar. Image resolution of the tissue image
file to load, either  \code{"lowres"} or \code{"hires"}.}
}
\value{
A \code{SpatialExperiment} object.
}
\description{
Load a Visium feature-barcode matrix and image as \code{SpatialExperiment}
}
\details{
The function assumes the standard Space Ranger directory structure:

\preformatted{
<visium_dir>/<visium_sample>/outs/
}
}
\examples{
\dontrun{
spex <- load_spex(
  visium_dir = "/path/to/visium",
  visium_sample = "sample_1",
  matrix_name = "filtered",
  matrix_file_format = "h5",
  image_resolution = "lowres"
)
}

}
\seealso{
\code{\link[SpatialExperiment]{read10xVisium}}
}
