% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tp.R
\name{load_tp}
\alias{load_tp}
\title{Load Visium tissue positions}
\usage{
load_tp(
  visium_dir,
  visium_sample,
  tp_file_name = "tissue_positions.csv",
  L_capture_area = NULL,
  check_values = TRUE
)
}
\arguments{
\item{visium_dir}{Character scalar. Base directory containing the Visium
sample directories.}

\item{visium_sample}{Character scalar. Name or identifier of the Visium
sample directory.}

\item{tp_file_name}{Character scalar. Name of the tissue-position file in
the sample's \file{outs/spatial} directory. Defaults to
\code{"tissue_positions.csv"}.}

\item{L_capture_area}{Positive numeric scalar defining the side lengths
of the squared capture-area. Supported values are \code{6500} and
\code{11000}. Defaults to \code{NULL}.}

\item{check_values}{Logical scalar. Whether required values and coordinate
types should be validated and \code{NA} values should be excluded.
Defaults to \code{TRUE}.}
}
\value{
Tissue positions as dataframe.
}
\description{
Load and optionally validate a Visium tissue-position CSV file.
}
\examples{
\dontrun{
tp <- load_tp(
  visium_dir = "/path/to/visium",
  visium_sample = "sample_1",
)
}

}
