% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixel_deviations_tp_json.R
\name{pixel_deviations}
\alias{pixel_deviations}
\title{Obtain axis-specific pixel deviations for the tissue positions file}
\usage{
pixel_deviations(tp_json)
}
\arguments{
\item{tp_json}{Data.frame with \code{array_row}, \code{array_col},\code{x},\code{y},
\code{pxl_row_in_fullres}, \code{pxl_col_in_fullres} columns.}
}
\value{
A named list with pixel deviation information:
\decribe{
\item{\code{lm_array}}{List object with linear models fitting the
\code{pxl_row_in_fullres} and the \code{pixel_col_in_fullres} coordinates to the
array grid.}
\item{\code{lm_slide}}{List object with linear models fitting the
\code{pxl_row_in_fullres} and the \code{pixel_col_in_fullres} coordinates to the
slide coordinates.}
\item{\code{residual_df}}{\code{tp_json} data.frame extended by 4 columns, namely
\code{dev_pxl_row}, \code{dev_pxl_col}, \code{dev_pxl_x}, and \code{dev_pxl_y}, comprising the
residuals for each linear model.}
\item{\code{residual_summary}}{Summary of the \code{residual_df}.}
}
}
\description{
Obtain axis-specific pixel deviations for the tissue positions file
}
\keyword{internal}
