% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{set_check_fun}
\alias{set_check_fun}
\title{set assertion checking functions}
\usage{
set_check_fun(...)
}
\arguments{
\item{...}{assertion checking function definitions, the name should be an
unglue pattern and the value the body of the function.}
}
\description{
set assertion checking functions
}
