% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_conversion.R
\name{spatial_conversion}
\alias{spatial_conversion}
\title{Convert Visium array and pixel coordinates to physical coordinates}
\usage{
spatial_conversion(
  aligned_tp,
  spex,
  L_capture_area,
  spatial_geometry = "hex100",
  image_resolution = c("lowres", "hires")
)
}
\arguments{
\item{aligned_tp}{Data frame containing all Space Ranger tissue-positions
with \code{array_row}, \code{array_col}, \code{pxl_row_in_fullres}, and
\code{pxl_col_in_fullres} columns, and aligned array (\code{array_x},
\code{array_y}) and pixel positions (\code{pixels_x}, \code{pixels_y}),
as returned by \code{check_fiducial_alignment()}.}

\item{spex}{SpatialFeature Object as returned by \code{load_spex()}.}

\item{L_capture_area}{Positive numeric scalar defining the side lengths
of the squared capture-area. Supported values are \code{6500} and
\code{11000}.}

\item{spatial_geometry}{Character Scalar. Defaults to \code{"hex100"},
representing the best supported spatial geometry settings with a nearest
neighbor spot distance of 100µm in a hexagonal grid.
Given the limited space of the capture area, exceeded by the default spatial
geometry grid \code{hex100}, we also support \code{k0}, corresponding to zero
distance between the spot's edges and the capture area edges, while
ensuring the containment of all spots.}

\item{image_resolution}{Character scalar. Image resolution to load, either
\code{"lowres"} or \code{"hires"}.}
}
\value{
A \code{SpatialFeatureExperiment} containing:
\describe{
\item{spatial array coordinates}{Spot-center coordinates transformed to the
requested spatial coordinate system.}
\item{\code{spatial_array}}{Buffered \code{sf} spot polygons in the spatial
array coordinate system.}
\item{spatial pixel coordinates}{Pixel coordinates transformed to the
requested spatial coordinate system.}
\item{\code{spatial_pixels}}{Buffered \code{sf} Pixel polygons in the
spatial pixel coordinate system.}
\item{transformed image}{The selected image with an extent aligned to the
spatial coordinates.}
\item{metadata}{Image and spot bounding boxes, scaling factors, the
original upper image boundary, and the spatial-unit label.}
}
}
\description{
Reconstruct physical coordinates for Visium array positions and transform
full-resolution image-pixel coordinates into the same coordinate system.
}
\examples{
\dontrun{
# convert spots spatially aligned tissue position & image data
for 6.5mm array (L_capture_area = 6500) into spatially 
aligned sfe object   
sfe_6500 <- spatial_conversion(
  aligned_tp = tp_aligned,
  spex = spex,
  L_capture_area = 6500,
  image_resolution="hires"
)
}


}
\seealso{
\code{\link[SpatialFeatureExperiment]{toSpatialFeatureExperiment}}
}
